feat(db): add car
add configuration

+++ src/database/Models/Session.cs
@@ -1,11 +1,12 @@
 ﻿
 using System;
+using System.Numerics;
 
 namespace SoborniyProject.database.Models
 {
     public class Session
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Key { get; set; }
         public int Status { get; set; }
         public string TotalTime { get; set; }
